+++ pytorch_lightning/overrides/base.py
@@ -14,6 +14,8 @@
 from typing import Any
 
 import torch
+from torch.nn import DataParallel
+from torch.nn.parallel import DistributedDataParallel
 
 from pytorch_lightning.core.lightning import LightningModule
 from pytorch_lightning.trainer.states import RunningStage
@@ -61,3 +63,12 @@ def warn_if_output_is_none(output: Any, method_name: str) -> None:
     """ Warns user about which method returned None. """
     if output is None:
         warning_cache.warn(f'Your {method_name} returned None. Did you forget to return an output?')
+
+
+def unwrap_lightning_module(wrapped_model) -> LightningModule:
+    model = wrapped_model
+    if isinstance(model, (DistributedDataParallel, DataParallel)):
+        model = model.module
+    if isinstance(model, _LightningModuleWrapperBase):
+        model = model.module
+    return model