+++ pytorch_lightning/plugins/training_type/ddp.py
@@ -20,10 +20,11 @@ from typing import Any, Dict, Optional, Union
 import numpy as np
 import torch
 import torch.distributed as torch_distrib
+from torch.nn.parallel.distributed import DistributedDataParallel
 
 from pytorch_lightning import _logger as log
 from pytorch_lightning.distributed import LightningDistributed
-from pytorch_lightning.overrides.data_parallel import LightningDistributedDataParallel
+from pytorch_lightning.overrides import LightningDistributedModule
 from pytorch_lightning.plugins.environments.cluster_environment import ClusterEnvironment
 from pytorch_lightning.plugins.training_type.parallel import ParallelPlugin
 from pytorch_lightning.utilities import _HYDRA_AVAILABLE
@@ -74,13 +75,6 @@ class DDPPlugin(ParallelPlugin):
     @property
     def root_device(self):
         return self.parallel_devices[self.local_rank]
-
-    @property
-    def lightning_module(self):
-        # the model may not be wrapped with DistributedDataParallel if calling this too early
-        # fixme: uncomment when this class will actually be used
-        # return unwrap_lightning_module(self._model)
-        pass
 
     @property
     def distributed_sampler_kwargs(self):
@@ -184,10 +178,8 @@ class DDPPlugin(ParallelPlugin):
         self.world_size = self.num_nodes * self.num_processes
 
     def configure_ddp(self):
-        # if unset, default `find_unused_parameters` `True`
-        self._ddp_kwargs["find_unused_parameters"] = self._ddp_kwargs.get("find_unused_parameters", True)
-        self._model = LightningDistributedDataParallel(
-            self.model,
+        self._model = DistributedDataParallel(
+            LightningDistributedModule(self.model),
             device_ids=self.determine_ddp_device_ids(),
             **self._ddp_kwargs,
         )
@@ -270,3 +262,12 @@ class DDPPlugin(ParallelPlugin):
         if isinstance(output, torch.Tensor):
             output = sync_ddp_if_available(output, group, reduce_op)
         return output
+
+    def training_step(self, *args, **kwargs):
+        return self.model(*args, **kwargs)
+
+    def validation_step(self, *args, **kwargs):
+        return self.model(*args, **kwargs)
+
+    def test_step(self, *args, **kwargs):
+        return self.model(*args, **kwargs)